requirements.txt
absl-py==2.1.0
access==1.1.9
aesara==2.9.3
affine==2.4.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
biopython==1.83
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clarabel==0.6.0
click==8.1.7
clickhouse-connect==0.7.8
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
cons==0.4.6
contourpy==1.2.0
coverage==7.4.0
cvxpy==1.4.1
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
dnspython==2.4.2
ecos==2.0.12
esda==2.5.1
ete3==3.1.3
etuples==0.3.9
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fiona==1.9.5
fonttools==4.47.2
fqdn==1.5.1
geopandas==0.14.2
giddy==2.3.5
idna==3.6
imageio==2.33.1
inequality==1.0.1
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.29.0
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
iteround==1.0.4
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy_loader==0.3
libpysal==4.9.2
llvmlite==0.41.1
logical-unification==0.4.6
mapclassify==2.6.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mgwr==2.2.1
miniKanren==1.0.3
mistune==3.0.2
momepy==0.7.0
mpmath==1.3.0
multipledispatch==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.3
opencv-python==4.9.0.80
orjson==3.9.12
ortools==9.7.2996
osqp==0.6.3
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
Pint==0.23
platformdirs==4.1.0
pluggy==1.3.0
pointpats==2.4.0
polars==0.20.22
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
pyarrow==16.0.0
pybind11==2.11.1
pycparser==2.21
Pygments==2.17.2
pymongo==4.6.1
pymunk==6.6.0
pyparsing==3.1.1
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.12; sys_platform == 'darwin'
PyQt5-Qt5==5.15.2; sys_platform != 'darwin'
PyQt5-sip==12.13.0
pyrsistent==0.20.0
pysal==23.7
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post0
qtconsole==5.5.1
QtPy==2.4.1
quantecon==0.7.1
rasterio==1.3.9
rasterstats==0.19.0
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Rtree==1.1.0
ruff==0.4.4
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.1
segregation==2.5
Send2Trash==1.8.2
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
spaghetti==1.7.4
spglm==1.1.0
spint==1.0.7
splot==1.1.5.post1
spopt==0.6.0
spreg==1.4.2
spvcm==0.3.0
stack-data==0.6.3
statsmodels==0.14.1
stochastic-arrow==1.0.0
swiglpk==5.0.10
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tobler==0.11.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
Unum==4.2.1
uri-template==1.3.0
urllib3==2.1.0
vivarium-core==1.6.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
